refactor: #1057 move jobs done label from logic to components
