+++ mw-webapp/src/component/label/Label.module.scss
