+++ mw-webapp/src/logic/wayPage/labels/LabelLine.tsx
@@ -3,11 +3,11 @@ import {observer} from "mobx-react-lite";
 import {Confirm} from "src/component/confirm/Confirm";
 import {EditableText} from "src/component/editableText/EditableText";
 import {HorizontalContainer} from "src/component/horizontalContainer/HorizontalContainer";
+import {LabelItem} from "src/component/label/Label";
 import {PositionTooltip} from "src/component/tooltip/PositionTooltip";
 import {Tooltip} from "src/component/tooltip/Tooltip";
 import {LabelDAL} from "src/dataAccessLogic/LabelDAL";
 import {languageStore} from "src/globalStore/LanguageStore";
-import {LabelItem} from "src/logic/wayPage/labels/label/Label";
 import {Label} from "src/model/businessModel/Label";
 import {LanguageService} from "src/service/LanguageService";
 import {debounce} from "src/utils/debounce";
