+++ mw-webapp/src/logic/wayPage/reportsTable/modalContentJobTags/ModalContentJobTags.tsx
@@ -4,10 +4,10 @@ import {observer} from "mobx-react-lite";
 import {Button, ButtonType} from "src/component/button/Button";
 import {Checkbox} from "src/component/checkbox/Checkbox";
 import {HorizontalContainer} from "src/component/horizontalContainer/HorizontalContainer";
+import {LabelItem} from "src/component/label/Label";
 import {PositionTooltip} from "src/component/tooltip/PositionTooltip";
 import {Tooltip} from "src/component/tooltip/Tooltip";
 import {languageStore} from "src/globalStore/LanguageStore";
-import {LabelItem} from "src/logic/wayPage/labels/label/Label";
 import {Label} from "src/model/businessModel/Label";
 import {LanguageService} from "src/service/LanguageService";
 import {KeySymbols} from "src/utils/KeySymbols";
